Add logging and uuids to the resource specification
I started changing the readme and things got out of hand.

The one *true* hell I remembered from debugging authorization
rules was determining what the relevant rule that was triggering
the decision was. We actually *don't* have a unique way to identify
what rules we're talking about in a system right now, which makes
debugging a lot harder.

This adds a configurable logger with debug settings for debugging
auth rules.